change quickstart name to stnapi; removes ordering in models
