+++ stnservices/stnapi/admin.py
