+++ stnservices/stnapi/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class QuickstartConfig(AppConfig):
+class StnAPIConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'quickstart'
+    name = 'stnapi'
